Cursor récupération position début

+++ PhotoGimp/Vertex.cpp
@@ -1,6 +1,6 @@
 #include "Vertex.h"
 
-Vertex::Vertex(float x, float y, float r, float g, float b)
+Vertex::Vertex(double x, double y, float r, float g, float b)
 	:x(x), y(y), r(r), g(g), b(b)
 {
 }
