+++ PhotoGimp/Vertex.h
@@ -2,8 +2,8 @@
 
 struct Vertex
 {
-	float x, y;
+	double x, y;
 	float r, g, b;
 
-	Vertex(float x, float y, float r, float g, float b);
+	Vertex(double x, double y, float r, float g, float b);
 };
